requirements.txt
flask
python-dotenv
google-generativeai
python-dotenv